Add first slack bot: Toggl-Weekly
Toggl-Weekly os a slackbot running on blockspring. It let's me add a
slackbot without a server and it's easily sharable so others can use it
too which is awesome.

What it does
you get your current weekly hourly report. So how many hours have you
worked. Might be extended shortly but for now just the hours is fine.
You can get the hours if you type in: /blockspring toggl-weekly